// packages/core/src/errors/enums.ts
export enum ErrorCodes {
  // 插件错误
  PLUGIN_ERROR = 'PLUGIN_ERROR',
  INVALID_PLUGIN = 'INVALID_PLUGIN',
  PLUGIN_HOOK_ERROR = 'PLUGIN_HOOK_ERROR',
  PLUGIN_LOADER_ERROR = 'PLUGIN_LOADER_ERROR',

  // 钩子错误
  HOOK_ERROR = 'HOOK_ERROR',
  INVALID_HOOK = 'INVALID_HOOK',

  // 加载器错误
  INVALID_LOADER = 'INVALID_LOADER',
  AUTH_LOADER_ERROR = 'AUTH_LOADER_ERROR',
  PAGE_LOADER_ERROR = 'PAGE_LOADER_ERROR'
}